requirements.txt
asgiref==3.2.3
Django==3.0.3
django-jalali-date==0.3.0
django-rest-auth==0.9.5
django-social-share==1.3.2
django-taggit==1.2.0
djangorestframework==3.11.0
intervaltree==3.0.2
jdatetime==3.6.2
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sortedcontainers==2.1.0
sqlparse==0.3.0
tag==0.3.3
